move prisma specific error
