+++ packages/jobs/src/adapters/PrismaAdapter/errors.ts
@@ -1,0 +1,8 @@
+import { RedwoodJobError } from '../../errors'
+
+// Thrown when a given model name isn't actually available in the PrismaClient
+export class ModelNameError extends RedwoodJobError {
+  constructor(name: string) {
+    super(`Model \`${name}\` not found in PrismaClient`)
+  }
+}
